+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   * `isSubsetOf` now assumes inputs have no duplicates
   * `isSubsetOf` third boolean arg to check if proper subset no longer exists
   * `isProperSubsetOf` now exported - and assumes no duplicates
+  * `compare` now split into `compare` and `compareBy`
 
 0.1.7 / 2015-11-15
 ==================